extract getting list of git files to its own module
