+++ packages/hasher/src/hashOfFiles.ts
@@ -1,8 +1,7 @@
 import crypto from "crypto";
 import path from "path";
-import globby from "globby";
 import fs from "fs-extra";
-import findUp from "find-up";
+import { getListOfGitFiles } from "./gitFiles";
 
 import { hashStrings } from "./helpers";
 
@@ -12,36 +11,14 @@ const LF = "\n";
 export async function generateHashOfFiles(
   packageRoot: string
 ): Promise<string> {
-  const nearestGitFolder = await findUp(".git", {
-    cwd: packageRoot,
-    type: "directory"
-  });
-
-  if (nearestGitFolder === undefined) {
-    throw new Error(
-      `It does not seem that this package is in a git repo: ${packageRoot}`
-    );
-  }
-
-  const repoRoot = path.dirname(nearestGitFolder);
-
-  // If the package is a git repo by itself then the search pattern is all the files
-  const relativePackagePath = path.relative(repoRoot, packageRoot) || "**/*";
-
-  // Note: globby does not support objectMode
-  const files = await globby(relativePackagePath, {
-    cwd: repoRoot,
-    gitignore: true
-  });
-
-  files.sort((a, b) => a.localeCompare(b));
+  const files = await getListOfGitFiles(packageRoot);
 
   const hashes = await Promise.all(
     files.map(async file => {
       const hasher = crypto.createHash("sha1");
       hasher.update(file);
 
-      const fileBuffer = await fs.readFile(path.join(repoRoot, file));
+      const fileBuffer = await fs.readFile(path.join(packageRoot, file));
       const data = fileBuffer.toString().replace(newline, LF);
       hasher.update(data);
